requirements.txt
gevent
nose
dnspython

